The Find dialog now displays the number of matches during find-as-you-type.
Initiating a search now disables most views, because many
more of them can initiate searches with find-as-you-type.

Using Cancel on the Find dialog now restores the results
from the most recent search in the history.  This has the
effect of undoing any damage to search results caused by
typing while the dialog was open.

Other minor robustness improvements to the Find dialog.