Finally found a reasonable way to fix Leopard's confusion over what the right application bundle is, so Python will no longer be the name of the application menu and preferences will no longer be horribly misdirected to Python's domain.
Starting with Leopard, the mini-interpreter from "Build Applet"
is used to trick Mac OS X into using the proper bundle.  This
is NOT preferred prior to Leopard, because it is non-portable:
it depends on a specific Python version, whereas prior to
Leopard a simple /usr/bin/python symbolic link will still work.
When Leopard's successor arrives, this will break!