It is now possible to reorder collections from the Preferences window drawer.  (This was always technically possible if low-level commands changed one of the "favorite-*" arrays when the program wasn't running.)
Although drag-and-drop is not yet supported, this
update was designed with drags in mind, and already
supports collection-relative positioning.

Fixed the initialization of context lists to now
use the domains explicitly read at startup time
(from disk), because domain name derivation can't
deal with a list of names that is not in numerical
order.