Major enhancement: enabled up to 256 colors, initially driven by XTerm sequences.  Also added all necessary changes to the Preferences user interface and Help documentation for this feature.
Attribute bits now indicate which custom colors 
are in use, and include the previous "ANSI color"
values.  To allow for future memory optimizations,
some bits have been redefined: the color indices
are now the upper 16 bits.  For safety, accessors
were first added to no longer directly manipulate
the various bits.

The public API for manipulating view colors has
been simplified, and internally views no longer
require all colors to be in the same "table".
There are new methods to access each class of
color separately.

This is still not a "complete" XTerm implementation,
so if an application tries to use other XTerm
features alongside the 256 colors, there may be
glitches.