Significantly changed the emulator data processing loop, to handle text translation according to the current input encoding. This required special care to continue to properly identify control sequences, so that translators may only "absorb" bytes that the terminal has no other use for.
Note that the renderer is still not Unicode
capable, but other uses for data (such as
drag and drop, or Copy) will correctly see
the translated values of any glyph.  Those
characters that are part of Mac Roman will
be rendered properly.

Removed some old APIs for sending CFStrings
directly to the emulator, that were unused
(but would have had major problems under the
new translation scheme anyway).