Wow...what a mistake it was to try using exceptions! Sampler data showed that throw-handling accounted for roughly 1/4 of the total compute time while scrolling even unstyled screenfuls of text.
All exceptions have been removed from the emulation
loop, and replaced with a "handled" flag that is
checked equivalently.  In addition, the parameters for
states in the determination and transition callbacks
are now a single reference to a pair, which shortens
the (commonly used) function signature and is also
a bit clearer in some situations.