Fix for #39, a long-standing (and very annoying!) bug with text selection and deselection in scrollback lines.
Part of the fix involved an enhancement to the iterator abstraction; it is
now possible to transparently move from a scrollback iterator into the
main screen, and vice-versa (which is arguably how it should have worked
to begin with).

Significantly enhanced the text search feature, since it was changing to
deal with terminal line offsets, anyway.  Originally written prior to the
CFString back-end, the search had been allocating a huge contiguous buffer
prior to every search, which was very costly!  It now searches each line's
raw string directly.  The direct search is generally faster.  It will,
however, make it harder to find words that wrap lines (still not handled).