Fixed a problem where capture files were not being closed properly; this could lead to many problems, but it would certainly prevent the same file name from being chosen for later captures.
In order to simplify cleanup, the Stream Capture provides
an auto-close option that is now used by the Terminal API.
This doesn't violate encapsulation, because the Session
module can now relinquish ownership of the file it opens,
in the same code that opens the file.  This is also much
simpler than something like a Terminal event handler.