Changed the behavior of Localization_ArrangeButtonArray() to allow each button to remain at ideal width, while being arranged (instead of forcing all the same width).  Therefore, this function no longer returns a width value, and callers had to change to ignore it.
Modified the default width of button-autosizing functions to
be more appropriate for Aqua.

Added an API to Generic Dialog that allows the title of the
default button to be changed.

Modified Generic Dialog so Localization_ArrangeButtonArray()
is used whenever buttons are changed (e.g. setting the title
or adding a 3rd button).