As a temporary measure, splitting teardown into two phases: one that is "safe" to re-enable, and one that is "unsafe" (and therefore still disabled).  This allows certain modules to once again tear down properly.
For some time, it has been unsafe to tear down any module
that might be depended upon by code in the destructor of a
custom HIObject (such as a terminal view), because Carbon
is capable of triggering these AFTER the application thinks
it has terminated.  Those modules are still not torn down
properly, and likely won't be until a transition to Cocoa
is complete.

Note that these API changes were technically needed by the
previous revision in order to compile...oops.