Added a class parameter to TerminalWindow_ReconfigureViewsInGroup(), and expanded the routine to copy settings to every terminal view *or* to every terminal screen buffer, depending on the class.
Used the modified API to finally fix Session-based new windows, so
that any associated Format, Terminal and/or Translation is correctly
copied.  (For example, this allows a user to choose a particular
terminal screen size whenever a certain Session is started.)  But,
note that the routines in the Terminal and Terminal View modules
that copy preferences are incomplete, so initial values won't be
quite right until those functions are flushed out.

Added the ability for SessionFactory_NewSessionArbitraryCommand() to
skip the now-implemented initialization from associated classes,
because it is sometimes called by other factory routines that will
already have handled this.

Discovered and fixed the arguments passed in to a new session spawned
from a macro (the session argument was missing, causing the workspace
index to be incorrectly interpreted as a nullptr value for a different
argument!).