Fixed session cloning (used by the "Duplicate Session" command) to create a terminal window when none is given, instead of dying with an assertion failure.  Also, this command now emits a beep if it fails for some reason.
Note that the cloner still doesn't retrieve preferences contexts
from the original window, so the new terminal window will not
yet copy features such as fonts, colors, and terminal size.
This should be fairly trivial, as the hard work of being able to
copy an entire context is already done; it's just a matter of
giving terminal windows a way to expose the preferences contexts
that represent their current settings.