The definition of a "word" is now defined entirely by Python code, and replaces the complicated C++ code for double-click highlighting. The initial implementation still breaks only on whitespace, but it will now be much easier to provide more elaborate options (such as breaks on certain punctuation characters).
Since this new callback is one of the most unusual yet (returning a
tuple and taking multiple arguments), some infrastructure also had
to be added to Quills to support this function call.

Minor cleanup to various Quills files: primarily to fix things like
copyright in comment headers, and format inconsistencies.