Added a forwarding mechanism to Commands_Executor so that some methods for validation and execution of menu commands can finally be implemented on other objects, like views and windows.  (Eventually, all such methods will be distributed like that, and Commands_Executor will go away.  The Commands_Executor only exists right now to deal with issues brought on by the Carbon/Cocoa hybrid.)
Since Commands_Executor is wired directly to menus (as it will be until
it goes away completely), the Cocoa responder chain doesn't actually see
methods on views and windows until Commands_Executor allows them to be
found.  All methods that support forwarding are therefore now marked
with a comment boilerplate, and share some code.  When Commands_Executor
is finally gone, the responder chain will simply find the methods on
windows and views automatically.

With this first commit, forwarding is being used for a couple of things
on the Cocoa terminal window (which is under test, and only available
from the debugging interface).  It is now possible to change the Format
of that window by using the usual menu commands; the forwarding of the
validators means that the commands are enabled whenever the new test
terminal window is in front!