Modified APIs that dealt with the concept of the current session or terminal window, so that it's possible to tell the difference between a terminal window that is actually focused, and one that is indirectly the target (e.g. because a floating window has the keyboard focus).  Updated every command handler, and any other code that depended on this, to decide exactly which target was necessary for correct behavior.
With this fix to commands and their validators, an odd problem
is finally fixed: sheets on terminal windows won't enter a state
where sometimes menu items are mostly-enabled (as if a terminal
were frontmost), and other times mostly-disabled (which would be
expected for the sheet).

Also, the current session no longer automatically looks at the
Session Info window, because few commands could do anything
useful with this, and some could be confused by it.  Now, each
command that knows how to handle the selected row in that window
is doing an explicit lookup to find it (when no other current
session is apparent).

Validators are now able to reliably find the terminal window
behind a sheet (say), so some commands now remain enabled after
a sheet opens.  For instance, you can now use macros, "nudge"
windows, suspend/resume, print the screen, and change certain
flags, even if a sheet is open on the terminal window.