Substantially tweaked the logic used when new terminal windows are displayed, so that they choose more reasonable workspaces of tabs.  Details are in the release notes.
Cleaned up the handling of workspaces in general in the
Session Factory, by adding a createWorkspace() helper and
updating various bits of code to use it.

Added a dynamic item to the File menu to show the user
that the Option key has a special effect on the workspace
of a new session.  Also took the opportunity to hide the
Shell option by default and make Log-In Shell prominent.