Rewrote portions of the code that handles margins, padding and layout of subviews in the Terminal View module (and to some extent the Terminal Window).  This has fixed various minor and major performance and rendering bugs.
The most significant difference is in Terminal Windows,
which now enforce a sane one-way control scheme: that is,
a window will never again adjust its size because one of
its views or screen buffers changed.  Window size changes
are only possible from the window level, and changes to
view sizes are made directly from the window.  This is
only natural, but the previous scheme seemed necessary
because it was possible for things like sheets to make a
change in a screen or a view that the window didn't find
out about.  A window is now able to monitor the changes
that a sheet makes, and it adjusts views appropriately.

Also, it's impossible to implement a split-pane without
making this change.  If a window were to immediately
respond to its view changes, then two different view
panes would trigger two incompatible window updates at
the same time.