Added storage for G2 and G3 character sets defined in the VT200 series, and support for basic sequences to use them. Per-character (single shift) and right-hand-side lock shifts aren't handled yet however.
Added valid states for all documented character sets, so
that the emulator will accept them.  But, so far most of
the character sets are not supported, and using them will
display console warning messages.

Fixed a problem with the code that's meant to skip VT100
no-op characters (codes 0 and 127).  It failed to skip the
characters, and even allowed them to echo to the screen.
This was potentially very serious if the byte immediately
following the character was important, such as an escape,
because the important character could be lost.