Implemented the VT220's "DECSCA" (select character attributes) sequence, which required a new flag in terminal text attributes to indicate can-erase/cannot-erase.
Also implemented "DECSED" and "DECSEL", which use the new erase
attribute that is managed by "DECSCA".  This allows parts of
the screen to be protected by applications during an erase.

To handle selective erase properly, bufferErase...() routines
needed a new parameter to indicate which steps to perform.
Also took the opportunity to clean up the code a bit and make
it more uniform (there's an opportunity to compress out some of
the redundancy, but for now it's OK).

In the process of looking at erase routines, found and fixed a
bug in how erases worked when going from the left edge up to
the cursor.  The correct buffer text was being cleared but the
wrong window area was updated, which meant that the user didn't
necessarily see the erase take effect immediately.  This could
be triggered by programs like "lynx" that rely on that type of
partial erase when rendering screens of text.

Moved the 'ICH' and 'ECH' implementations from XTerm into VT220
(where they belong), which allows insert/erase to work even if
XTerm features are not enabled.

Fixed console warnings in the Terminal module to properly call
Console_Warning() as a wrapper, though most of them were already
commented-out anyway.  Also moved the unexpected-input messages
into if-statements that check the appropriate debug flag.