XTerm is now its own terminal type, and it has been added to the Terminals pane as a Base Emulator option. Selecting the option will automatically set a sensible Identity string and enable various tweaks, although all of those things can still change separately.  The new XTerm base emulator is not yet the default because it is not complete (in some situations applications will still perform better if they use VT220 or less).
The Terminal module has been refactored somewhat so
that XTerm "core" features are now in a separate
class.  The core class is not capable of being its
own emulator; it is relied upon by either the class
that implements XTerm, or any terminal that has
installed XTerm-related emulation tweaks.  Comments
on this feature have been extended to describe how
the so-called pre-callbacks are intended to work.

Updated MacTelnet Help to have accurate descriptions
of the current emulators.