Fixed a problem in the Custom Key Sequences sheet (which relies on the Control Keys palette to operate): keys would be handled correctly by the sheet, but would also be forwarded to the session underneath.  Now the session tracks the sheet and refuses to handle any key requests until the sheet is closed.
This was happening because the Control Keys palette really
does forward its clicks to multiple locations right now.
This is an artifact of the incomplete transition to Cocoa;
ultimately it should only be necessary to rely on Cocoa to
forward events to the first responder.  Until then, the
solution is a bit inelegant.