Modified the appearance of alert messages; there is now slightly more whitespace, and titled alerts have soft bluish-gray Helvetica text.  The window title no longer shows the name of the application.
Changed the code of the Alert Messages module slightly so
that it depends a little more on NIB layout and less on
its assumptions about that layout.

Fixed a bug where in right-to-left localization mode the
alert's smaller application icon would not be moved to
the other side of the window.

Updated the internals of the Localization module somewhat:
reformatted some code, updated a few API calls and removed
a number of obsolete APIs (such as those dealing with the
very old concept of the Dialog Manager and dialog items).