Implemented the "back color erase" feature that is depended upon by a significant number of programs (especially important in XTerm mode).  It is available as a tweak so that it can be used in lesser terminals.
This is an unusual feature in that the terminal must now keep
track of changes to the background color attribute.  The most
recent setting could be copied many times to other regions.

Since this touched a series of routines that deal with insertion
and deletion of characters and lines, the opportunity was taken
to make those interfaces slightly more robust.  Now, character
routines require "rules" for dealing with attributes, just as
their line equivalents have required.  A new type of rule allows
the caller to copy the most recent background color to the new
blank regions.

Reimplemented the mechanism for handling terminal variants, since
there are still only a few of them and they are checked far too
often to pay the cost of a set lookup.  Bits are used instead.