Updated all code that uses a name to find an existing preferences collection, so that Preferences_IsContextNameInUse() is called before Preferences_NewContextFromFavorites().  This prevents an invalid name from saving a new, blank collection with that name.
For now, this code only prints a console warning if it can't find
the context it was looking for.  In the future it will be better
to make context renames smarter, e.g. so they can automatically
patch preferences that contain references to the old name.  The
user should be free to rename a collection without worrying about
breaking other references to that collection.