Updated the Alert Messages module so that modeless alerts (background notifications) use Cocoa.  Other alerts are still Carbon-based for historical reasons that are not yet easily avoided.
As part of the change, modeless alerts now require
the Alert_NewModeless() call, and the old routine
Alert_MakeModeless() has been removed.  This is not
only cleaner, but it was necessary to ensure that
certain state in the Cocoa window could be set
correctly.

Fixed a minor problem with the placement of the
miniature application icon in Carbon-based alerts.

Since certain Cocoa Bindings code is now used in
more than one place, some generic routines from the
Server Browser have been moved into a new module,
Cocoa Extensions.  Both Server Browser bindings and
the new Alert Messages bindings use the relocated
code.  Also, fixed a problem with this binding code
that caused the auto-key-update mechanism to use
an invalid technique to search for class methods.