Converted all remaining references to the old application name, except those that are there because they deal with legacy stuff. This includes finally renaming the main bundle and its scripts, and references to that bundle in user interfaces and documents.
Also fixed a bug in the front-end script that would allow the
Preferences Converter to run even if the user preference version
was zero (i.e. brand new preferences).  This prevents possible
data corruption, but it also reveals that the program does not
deal gracefully with unexpected values such as large integers.