CSI parameter handling has been redone so that it is integrated into the rest of the terminal state machine. This creates different complexity, not necessarily less, but it should avoid any obscure corner cases.  The actual motivation for doing this was to eliminate the look-ahead scan that the previous parameter code was doing.  By removing an alternate path through the byte stream, there will ideally only ever be one place to explicitly decode UTF-8 bytes in the near future (for sessions using that).
Also fixed the parameter terminators for 'ICH' and 'ECH'
so that they will work in VT220-only terminals (which
they should; previously they were confined to XTerm).

It should be noted that the state machine code is still a
bit bigger and weirder than it ought to be, and in some
ways this change only compounds those general problems.
Future commits should aim to improve all of this.