Migrated application-modal alerts to Cocoa; they now share much of the same code as the modeless notification windows. This required changes to code that constructs *sheets* (even though they are still Carbon-based) because sheets now have a separate construction routine.  (Some code changes were less trivial because the construction routine now requires more parameters to be known up front.)
Fixed the code that adjusts the alert window size and view
layout when the title or help text is not provided.  Previously
the dialog text could be put in the wrong position.  (This was
also affecting notification windows, but no existing modeless
alerts were configured in a way that would show the problem.)

Fixed the tab order for buttons in the new Cocoa-based alerts
(both notification windows and application-modal alerts).

Added a Cocoa version of the button auto-size helper in the
Localization module, and as such changed the module into an
Objective-C++ file.