Changed the previous Option-resize preference-swapping behavior to now require a Control-resize.  (As of Mac OS X Lion, other modifier keys had their own default effects on window resizing so Control is the only real, er, option.)  Updated documentation to match.
Unfortunately the Option key must still be used to alter the
behavior of Full Screen.  If the Control key were required, the
mode couldn't be triggered when starting from the toolbar (as
a contextual menu would open).  Also, the default menu item key
equivalent for "Enter Full Screen" requires the Control key.

Note that the grow-box-click event handler has been removed
because on Mac OS X Lion it was insufficient for dynamically
changing the behavior of a resize.