Expanded the "Function Keys" palette to 48 keys, and added Map menu items to choose between 4 possible layouts.  This allows MacTerm to support the original VT220 mode as well as some popular variants from XTerm and "rxvt".
Added low-level Preferences settings for the new layouts so
that they can be easily saved and restored.

Significantly extended the infrastructure in the Terminal
and Session modules that dealt with function keys, so that
all of the new key codes from the 4 supported layouts can be
transmitted.  This was tedious to build but it should not
need to be changed again.

To make the layout more convenient to set, a pop-up menu is
now seemingly embedded in the title bar of the Function Keys
palette.  In reality the menu is in a separate child window
that is translucent and kept in sync with the window title
region of the parent.

An animation effect has been added that can "balloon" the
font size of a keypad button momentarily.  This is used when
the function key layout is changed, to show any keys that
are not really function keys anymore (such as the "do" key).

Took the opportunity to fix a few Interface Builder warnings
for floating windows, primarily by removing auto-save name
settings (that are not read on Panther) and replacing them
with "windowFrameAutosaveName" methods in various controller
classes.

Updated MacTerm Help to describe the new layouts in detail,
and to describe the low-level setting that was added.