Added a new Cocoa Animation module, that provides some simple APIs to create more interesting window animations on ANY version of Mac OS X...
Currently, there are only two exposed APIs (to show a
window-remove animation or window-duplicate animation),
though the underlying class is a bit more flexible with
some options for time distribution and effects like a
fade-in or fade-out.  The APIs also provide a quick way
to cut down animation complexity; callers are expected
to set this flag when it is apparent that the user is
running on a slower machine (right now, a relatively
simple OS version check is performed to decide this).

The animation makes itself independent of a window by
taking a snapshot of the window image and rendering the
image in a new window that is completely owned by the
animation.  This is particularly valuable for window-close
effects, since the original window can still be killed
immediately while appearing to exist for awhile longer.

Updated the "Duplicate Session" command to use the new
animation module.  This required a way to spawn a window
without any staggering, because the animation is meant to
simulate a new window spawning out of an existing one.

Updated the closing of terminal windows to use the new
animation module instead of simple zoom rectangles.