Added the ability to animate windows between different sizes of frames, and modified the standard window-close animation to use this.  The image windows now scale their content "to fit" so that the intermediate frames do not contain any white edges.
It is now possible to set the overall delay of the
animation, although this is a rough estimate at best
due to other factors in delay.  Also improved the
algorithms (such as fade-out) so that they can properly
respect this maximum delay.

Added a new animation for hiding windows, which takes
a target rectangle as a parameter.

Simplified the Cocoa Animation APIs so that optimizations
such as frame-dropping are selected automatically,
internally, instead of being exposed as a parameter.

Fixed a problem where a "duplicate" animation (which
moves windows offscreen) could leave windows offscreen;
now, multiple measures are taken to ensure that the
chosen final frame is definitely on the screen and that
the window uses this frame even if animation fails.