Timers are now created in a Session to manage the pasting of each line of text (when not choosing to join all text together as a single line) so that a user-specified delay can be imposed after each line.
Added a low-level preference setting to manage the
delay value.  The factory default is very small
(almost instantaneous).

The splitting of pasted lines actually has several
additional advantages.  Each line is now followed
by the proper new-line sequence for the target
session.  When lines are split from their source,
additional types of line separators are allowed
(such as those in Unicode) so there is a better
chance that the text will be handled correctly.
And, by not trying to paste large chunks of text
the terminal is less likely to overwhelm a process
and cause a hang.

Fixed a timer installation problem on Panther.
According to Apple documentation timers cannot be
installed with a start time of zero on Panther, so
they now just use really small numbers instead.
This should fix timer-dependent routines on that
OS (for example, Clipboard window updates).

Removed some Clipboard APIs that were too closely
coupled to other modules anyway, and made their
callers just directly perform the same actions.