The "Restart Session" command has been enhanced so that it can run at any time.  If it is invoked while a process is still running, it kills the previous process and then starts a new one after a short delay (the delay is needed to successfully restart processes such as shells).
The alert that appears prior to closing a window has been
enhanced so that a modified version of it appears prior to
a restart of a long-running session.

The behavior when setting the state of a session has been
cleaned up somewhat, but the restart-running-process
enhancement has also added some complexity.  Basically,
it is now possible for a process to die as part of the
restart procedure, and this type of exit must never kill
the terminal window.

Added a toolbar item for the "Restart Session" command,
and updated the default toolbar layout slightly to give
the new item a sensible placement.