Preferences window Sessions pane Data Flow tab has been improved: the Local Echo setting has been simplified to be a checkbox, there is a new field to edit the delay value for inserted lines, and the scroll speed slider is now functional (although nothing is yet reading the value of the preference that it saves).
Fixed the default value of the scroll delay preference to
be zero.  Compensated for that by making the low-level
preferences system refuse to honor any "large" values for
the setting; otherwise one day the terminal will start to
scroll very slowly and the user may not realize why.

The MacTerm Help content has been improved for the above
preferences as well as related features of the program.