Fixed the callback model and event handlers for the Find dialog to resolve cases where it was not being destroyed.
This required creating new utility routines to scan for
keyboard events that might activate buttons (default or
Cancel).  It turns out that if a key filter for a field
calls through to its next event handler and the current
key event is something like a Return or an Escape, the
filter routine will still be on the call stack when the
button command is activated.  Previously this meant that
an automatic dialog lock in the filter code would keep
the button handler from destroying the dialog.  Now, the
filter not only checks for possible button keys before
it locks the dialog, but it won't even try to call the
next handler if the event is targeting a button.