Adding Matt Gemmell's excellent MAAttachedWindow implementation to the project (along with appropriate credits in the About box). This will be used to implement a variety of pop-over windows over time.  A very minor tweak was required to make it also work on Mac OS X 10.3.
Some utility routines were added for setting up the style of a
pop-over window, which bumped the version of the owning library.
(In the future the library version should be removed from all
file headers to minimize changes in this kind of situation.)

Converted the Find interface into a Cocoa window (the first
window to use the pop-over style).  Minor changes to the Commands
module were required to ensure that Edit menu actions would reach
the Cocoa window when it is attached to a Carbon-based window.