Fixed logic of popover manager in the event that the parent window is a sheet and not a main window.
Popovers no longer close automatically if they have
an open sheet.  This is useful for things like form
validation errors for instance.

Fixed popovers to call "close" instead of "orderOut"
(and users of popovers to not release on close) so
that the hiding of a popover can be detected by a
close notification.