The Servers panel has been converted into a popover, which should make it easier for the user to understand its effects (previously the panel would float somewhere on the screen but now it points to the text field that is being changed).
As with other popovers, the UI for Servers has now been split
into a XIB that defines just the enclosing view (where a
separate class decides what kind of window to wrap it in).
This required a lot of code shuffling, though not too many
actual changes.

The needs of the Servers panel revealed some minor issues
with popovers that have been corrected in this commit and in
some other recent commits.  Basically, the window can resize
so the resize animation needs to be supported and the arrow
must be relocatable whenever the new size goes offscreen.
Also, the parent window is now forced to deactivate because
Cocoa popovers do not easily "click to remove" when the
parent is a Carbon document-class window like Preferences.