Import/export of ALL kinds of preference collections has been added to the user interface of the Preferences window as new menu commands in the drawer.  (This has been possible in the low-level code for some time, but it is finally exposed to the user.)  There is also a Python API for the import step so that the appropriate preference file extensions can be "registered" in Python code in the same way as other file types.
A new icon and file type description has been added so that
the Finder and other parts of the system realize that property
list files may be opened with MacTerm.

Although the user interface is technically only in the
Preferences window, Navigation Services is designed to open
files via Apple Events.  This means preference files can be
opened at any time and the user might not even be looking at
the Preferences window when the imports occur.  So, the event
handler forces the proper pane of the Preferences window to
be displayed whenever a file is imported.  The pane to show
is inferred from the kinds of settings in the imported file.

Some effort has been put into fixing the differences between
open-URL events and open-document events.when the targets are
local files.  While this needs more work, file opens should
now behave more consistently (e.g. dragging an icon to the
Dock or using Navigation Services sends an open-documents
event but double-clicking in the Finder sends open-URL; the
URL case now behaves like open-documents for local files).