Implemented Cocoa versions of every toolbar item, and added an experimental Cocoa-only floating toolbar window.  For now however the Carbon window toolbars remain in use and the Cocoa one is only accessible through the debug interface.
Due to a very odd design decision in the OS, NSToolbar is
not an NSView so it can't be inserted anywhere except where
Cocoa wants it to go.  It can't even go in custom windows!
Since long-term the terminal window will almost certainly
have a custom style, the only real way to use a standard
toolbar is to put it in some standard window.  So a global
toolbar has been built.

The global toolbar is a floating window with NSToolbar and
all standard items, customization, etc.  It is coupled with
code to pin the toolbar to either the top or bottom of any
display (whatever the user wants).  The toolbar fills the
entire horizontal width of its display and it becomes
translucent when it loses the focus.

This design has some advantages: it's a standard toolbar,
it can take up less screen space than the toolbars of
several open windows, and it leaves complete freedom to
redesign the terminal window style without losing toolbar
functionality.  On the other hand, the window is sometimes
in the way; code will have to be written to manage windows
around the toolbar (much like the OS keeps things away
from the Dock).