Significantly changed the mechanism for building version web pages. The parser for the mini-language is now more general, capable of reading any kind of section or loop reference.  The mechanism for establishing context (variables defined within a loop, etc.) has also been refined so that it is easier to see variable settings and dependencies.
Modified the "daily build" log format so that each year appears in
its own section (complete with a disclosure triangle to hide the
entire year's log entries and a link at the top of the page to jump
to that year's section).

Fixed domain names in template HTML files to use the "www" variant.

Fixed the "clean" rule in the GNUmakefile to remove all the files
that are actually generated.
 
Removed the "pad.plist" option, which hasn't been used in years.