Major improvements to the vector graphics canvas (almost ready for release but in this commit still guarded by a checkbox in the debugging interface):
- A Cocoa version of the graphics window is now implemented.
  The canvas renders all lines smoothly.

- The mechanism for copying graphics to the clipboard has
  been reimplemented, now supporting both bitmap (TIFF) and
  PDF copies of the data.  This allows applications that
  understand PDF to see the original drawing without any
  loss of quality or flexibility.

- It is now possible to print vector graphics windows!
  The printed image automatically uses black and white as
  the foreground and background respectively (even if the
  window has customized these colors), although any other
  colored lines are still printed.

There are still a few rough edges to iron out.  One is
that the "rename window" popover has to change to display
on top of the new Cocoa window.  Also, if a graphic is
relying too heavily on individual pixels (e.g. to make
its own fake font for text) the new engine can actually
make those fonts look *worse*; there may need to be an
option to dumb-down the rendering in certain situations.