The new Cocoa-based vector graphics canvas is now the default implementation; its capabilities are described in previous commits. Since this is no longer an experimental feature the relevant parts have been removed from the debugging interface.
A Vector Window module has been implemented so that the window
controller is easier to manage (e.g. from the Session module).

The Vector Interpreter object now has a proper reference type
instead of just an ID and it can be retained/released normally.
The window controller now uses the interpreter exclusively and
only indirectly uses the Vector Canvas through its main view
(this is pretty important because an interpreter is ultimately
the only thing that creates or deletes a canvas).

The Session module's TEK-related API has been reduced and instead
a new scheme is used to manage graphics internally.  The module
now tracks all TEK windows that a particular session opened but
still allows only one to be the current data target.  The new
Vector Window module provides an event monitoring system that the
Session uses to find out when windows close.