Vector canvases are now explicitly told the purpose of the lines they are asked to draw (text or graphics), and the canvas chooses to make graphics lines thicker than text lines. This change makes text more readable and it appears smoother.
Requests to draw single points have been tweaked so that they
can never be invisible; this is achieved by starting a new
path when a lone point is discovered and setting its line cap
style differently.  The possible downside is that this is not
currently detecting lone points in series; each time a point
is rendered a new path is forced after that point in order to
prevent the lone-point cap style from affecting non-points
that follow.