Cocoa version of Preferences window has been extended to include a Formats pane and one panel (so far) for editing the 16 standard colors.
The interface for the Cocoa version has slightly changed: it uses the
system's standard color box buttons, it provides "Inherit" checkboxes
for all colors individually (instead of a single box for normal/bold)
and it is named "Standard Colors" instead of "ANSI Colors".  Also,
like the new Translations pane, it integrates into the source list
that is provided by the Preferences window.