More additions to the Cocoa version of the Preferences window: implemented the "Formats" pane's "General" tab (for setting terminal font, size, etc.).
This uses the new Cocoa implementation of the Terminal View, which is
incomplete but still good enough to serve as a "Sample" pane.

The code for managing the "is inherited", "inherit enabled" and "preference
value" settings has been shuffled around in an attempt to be more reusable.
There is now a generic class for handling this triplet of properties for
any preference whose tag is a color, and another general class for any
preference whose tag is a CFStringRef.  Bindings have been changed to match.
This approach should make it MUCH easier to add "Inherit" checkboxes to
virtually all panels (whereas in the Carbon version this was such a pain
that it was avoided in most cases).